+++ purchase_order.go
@@ -17,20 +17,20 @@ type PurchaseOrder struct {
 	PaymentMethod     int32      `json:"paymentMethod"`
 	BillingSeries     string     `json:"billingSeries"`
 	Currency          int32      `json:"currency"`
-	CurrencyChange    float32    `json:"currencyChange"`
+	CurrencyChange    float64    `json:"currencyChange"`
 	BillingAddress    int32      `json:"billingAddress"`
 	ShippingAddress   int32      `json:"shippingAddress"`
 	LinesNumber       int16      `json:"linesNumber"`
 	InvoicedLines     int16      `json:"invoicedLines"`
 	DeliveryNoteLines int16      `json:"deliveryNoteLines"`
-	TotalProducts     float32    `json:"totalProducts"`
-	DiscountPercent   float32    `json:"discountPercent"`
-	FixDiscount       float32    `json:"fixDiscount"`
-	ShippingPrice     float32    `json:"shippingPrice"`
-	ShippingDiscount  float32    `json:"shippingDiscount"`
-	TotalWithDiscount float32    `json:"totalWithDiscount"`
-	VatAmount         float32    `json:"vatAmount"`
-	TotalAmount       float32    `json:"totalAmount"`
+	TotalProducts     float64    `json:"totalProducts"`
+	DiscountPercent   float64    `json:"discountPercent"`
+	FixDiscount       float64    `json:"fixDiscount"`
+	ShippingPrice     float64    `json:"shippingPrice"`
+	ShippingDiscount  float64    `json:"shippingDiscount"`
+	TotalWithDiscount float64    `json:"totalWithDiscount"`
+	VatAmount         float64    `json:"vatAmount"`
+	TotalAmount       float64    `json:"totalAmount"`
 	Description       string     `json:"description"`
 	Notes             string     `json:"notes"`
 	Off               bool       `json:"off"`
@@ -258,7 +258,7 @@ func (p *PurchaseOrder) deletePurchaseOrder() bool {
 
 // Adds a total amount to the order total. This function will subsctract from the total if the totalAmount is negative.
 // THIS FUNCTION DOES NOT OPEN A TRANSACTION.
-func addTotalProductsPurchaseOrder(orderId int64, totalAmount float32, vatPercent float32) bool {
+func addTotalProductsPurchaseOrder(orderId int64, totalAmount float64, vatPercent float64) bool {
 	sqlStatement := `UPDATE purchase_order SET total_products=total_products+$2,total_vat=total_vat+$3 WHERE id=$1`
 	_, err := db.Exec(sqlStatement, orderId, totalAmount, (totalAmount/100)*vatPercent)
 	if err != nil {
