+++ sales_invoice.go
@@ -14,16 +14,16 @@ type SalesInvoice struct {
 	PaymentMethod      int32     `json:"paymentMethod"`
 	BillingSeries      string    `json:"billingSeries"`
 	Currency           int32     `json:"currency"`
-	CurrencyChange     float32   `json:"currencyChange"`
+	CurrencyChange     float64   `json:"currencyChange"`
 	BillingAddress     int32     `json:"billingAddress"`
-	TotalProducts      float32   `json:"totalProducts"`
-	DiscountPercent    float32   `json:"discountPercent"`
-	FixDiscount        float32   `json:"fixDiscount"`
-	ShippingPrice      float32   `json:"shippingPrice"`
-	ShippingDiscount   float32   `json:"shippingDiscount"`
-	TotalWithDiscount  float32   `json:"totalWithDiscount"`
-	VatAmount          float32   `json:"vatAmount"`
-	TotalAmount        float32   `json:"totalAmount"`
+	TotalProducts      float64   `json:"totalProducts"`
+	DiscountPercent    float64   `json:"discountPercent"`
+	FixDiscount        float64   `json:"fixDiscount"`
+	ShippingPrice      float64   `json:"shippingPrice"`
+	ShippingDiscount   float64   `json:"shippingDiscount"`
+	TotalWithDiscount  float64   `json:"totalWithDiscount"`
+	VatAmount          float64   `json:"vatAmount"`
+	TotalAmount        float64   `json:"totalAmount"`
 	LinesNumber        int16     `json:"linesNumber"`
 	InvoiceNumber      int32     `json:"invoiceNumber"`
 	InvoiceName        string    `json:"invoiceName"`
@@ -263,7 +263,7 @@ func (i *SalesInvoice) deleteSalesInvoice() bool {
 
 // Adds a total amount to the invoice total. This function will subsctract from the total if the totalAmount is negative.
 // THIS FUNCTION DOES NOT OPEN A TRANSACTION.
-func addTotalProductsSalesInvoice(invoiceId int64, totalAmount float32, vatPercent float32) bool {
+func addTotalProductsSalesInvoice(invoiceId int64, totalAmount float64, vatPercent float64) bool {
 	sqlStatement := `UPDATE sales_invoice SET total_products = total_products + $2, vat_amount = vat_amount + $3 WHERE id = $1`
 	_, err := db.Exec(sqlStatement, invoiceId, totalAmount, (totalAmount/100)*vatPercent)
 	if err != nil {
